Added api.rs and basic API functions
api.rs is going to be an API wrapper for the RPS API.
Then this api will be used by the package management system to install files, update files, and look up file information.
I plan for the package management system to keep a local copy of the /api/list.

Changelog:
- Added api.rs
- Added basic API functions