Updated dependencies in Cargo.toml
Removed two dependencies:
- reqwest
- serde

These dependencies are no longer needed as the API module is being reworked. They will likely be added in the future, this is just a clean up for now.

Added one dependency:
- colored

Added colored to display colored text in the terminal. This is helpful for highlighting parts of abort messages, and highlighting the help message.